Combine vm files and import usertests from xv6-riscv
...I am not even attempting to make those tests pass yet, but it shall allow more surface area to be covered. Also, I discovered a regression in the createdelete test. Still working on that.

+++ lib/queue.c
@@ -18,7 +18,7 @@ struct queue_##T##_node { \
 struct queue_##T { \
   struct queue_##T##_node *front; \
   struct queue_##T##_node *rear; \
-  int size; \
+  size_t size; \
 }; \
 \
 struct queue_##T * \
